requirements.txt
Flask==0.9
Jinja2==2.6
Werkzeug==0.8.3
distribute==0.6.24
wsgiref==0.1.2
mysql-python
requests
simplejson
Flask-SQLAlchemy==0.15
configobj
